modify

+++ src/main/java/com/thread2/Producer/ProducerPartitioner.java
@@ -13,8 +13,8 @@ public class ProducerPartitioner implements Partitioner {
     @Override
     public int partition(String s, Object o, byte[] bytes, Object o1, byte[] bytes1, Cluster cluster) {
        String timestamp = o.toString().split("_")[0];  //o是消息，object类型要先转换为String类型
-       long time=Long.parseLong(timestamp);
-       System.out.println(time);
+       Double time=Double.parseDouble(timestamp);
+       System.out.println(time.longValue());
        int numPartition=cluster.partitionCountForTopic(s);   //s是topic
        return (int)(time % numPartition);
     }
